added to_dict/from_dict tests
